+++ src/app/(protected)/(admin)/(tabs)/profile.tsx
