+++ src/app/(protected)/(admin)/_layout.tsx
@@ -1,14 +1,5 @@
-import { Tabs } from "expo-router";
-import { RoleBasedTabLayout } from "@/widgets/bottom-tab-bar/ui/RoleBasedTabLayout";
+import { Stack } from "expo-router";
 
-export default function AdminTabLayout() {
-	return (
-		<RoleBasedTabLayout userType="ADMIN">
-			<Tabs.Screen name="home" />
-			<Tabs.Screen name="map" />
-			<Tabs.Screen name="dashboard" />
-			<Tabs.Screen name="chat" />
-			<Tabs.Screen name="profile" />
-		</RoleBasedTabLayout>
-	);
+export default function AdminLayout() {
+	return <Stack screenOptions={{ headerShown: false }} />;
 }
