+++ src/app/(protected)/(partner)/(tabs)/_layout.tsx
@@ -1,0 +1,14 @@
+import { Tabs } from "expo-router";
+import { RoleBasedTabLayout } from "@/widgets/bottom-tab-bar/ui/RoleBasedTabLayout";
+
+export default function PartnerTabLayout() {
+	return (
+		<RoleBasedTabLayout userType="PARTNER">
+			<Tabs.Screen name="home" />
+			<Tabs.Screen name="map" />
+			<Tabs.Screen name="dashboard" />
+			<Tabs.Screen name="chat" />
+			<Tabs.Screen name="profile" />
+		</RoleBasedTabLayout>
+	);
+}
