+++ src/app/(protected)/(partner)/_layout.tsx
@@ -1,14 +1,5 @@
-import { Tabs } from "expo-router";
-import { RoleBasedTabLayout } from "@/widgets/bottom-tab-bar/ui/RoleBasedTabLayout";
+import { Stack } from "expo-router";
 
-export default function PartnerTabLayout() {
-	return (
-		<RoleBasedTabLayout userType="PARTNER">
-			<Tabs.Screen name="home" />
-			<Tabs.Screen name="map" />
-			<Tabs.Screen name="dashboard" />
-			<Tabs.Screen name="chat" />
-			<Tabs.Screen name="profile" />
-		</RoleBasedTabLayout>
-	);
+export default function PartnerLayout() {
+	return <Stack screenOptions={{ headerShown: false }} />;
 }
