+++ src/app/(protected)/(student)/_layout.tsx
@@ -1,13 +1,5 @@
-import { Tabs } from "expo-router";
-import { RoleBasedTabLayout } from "@/widgets/bottom-tab-bar/ui/RoleBasedTabLayout";
+import { Stack } from "expo-router";
 
-export default function StudentTabLayout() {
-	return (
-		<RoleBasedTabLayout userType="STUDENT">
-			<Tabs.Screen name="home" />
-			<Tabs.Screen name="map" />
-			<Tabs.Screen name="suggestion" />
-			<Tabs.Screen name="profile" />
-		</RoleBasedTabLayout>
-	);
+export default function StudentLayout() {
+	return <Stack screenOptions={{ headerShown: false }} />;
 }
